Assets panel upgrade & more
 - Open folder context window
 - Create folder button
 - Inspector formatting
 - Scene load payload
 - Mesh view load only .fbx payload
 - Warnings reduced 6 (build only)
 - Icons switched to white

+++ Wiwa/src/Wiwa/utilities/render/primitives/Sphere.cpp
@@ -17,7 +17,7 @@ namespace Wiwa {
         // Generate vertices
         for (int i = 0; i <= stack_count; ++i)
         {
-            stackAngle = PI / 2 - i * stackStep;
+            stackAngle = (float)PI / 2 - i * stackStep;
             xy = radius * cosf(stackAngle);
             z = radius * sinf(stackAngle);
